MudSelectItem, MudNavLink: Remove MudBaseSelectItem
