+++ src/MudBlazor/Components/NavMenu/MudNavLink.razor
@@ -1,5 +1,5 @@
 ﻿@namespace MudBlazor
-@inherits MudBaseSelectItem
+@inherits MudComponentBase
 @using Microsoft.AspNetCore.Components.Routing
 
 <div @attributes="UserAttributes" class="@Classname" disabled="@Disabled" style="@Style">
