+++ src/MudBlazor/Components/NavMenu/MudNavLink.razor.cs
@@ -1,12 +1,13 @@
 ﻿using Microsoft.AspNetCore.Components;
 using Microsoft.AspNetCore.Components.Routing;
+using Microsoft.AspNetCore.Components.Web;
 using MudBlazor.Interfaces;
 using MudBlazor.Utilities;
 
 namespace MudBlazor
 {
 #nullable enable
-    public partial class MudNavLink : MudBaseSelectItem
+    public partial class MudNavLink : MudComponentBase
     {
         protected string Classname =>
             new CssBuilder("mud-nav-item")
@@ -35,6 +36,9 @@ namespace MudBlazor
         }
 
         protected int TabIndex => Disabled || NavigationContext is { Disabled: true } or { Expanded: false } ? -1 : 0;
+
+        [Inject]
+        private NavigationManager UriHelper { get; set; } = null!;
 
         [CascadingParameter]
         private INavigationEventReceiver? NavigationEventReceiver { get; set; }
@@ -71,6 +75,56 @@ namespace MudBlazor
         [Category(CategoryTypes.ComponentBase.Common)]
         public string ActiveClass { get; set; } = "active";
 
+        /// <summary>
+        /// Prevents the user from interacting with this item.
+        /// </summary>
+        [Parameter]
+        [Category(CategoryTypes.General.Behavior)]
+        public bool Disabled { get; set; }
+
+        /// <summary>
+        /// Shows a ripple effect when the user clicks the button.
+        /// </summary>
+        /// <remarks>
+        /// Defaults to <c>true</c>.
+        /// </remarks>
+        [Parameter]
+        [Category(CategoryTypes.General.Appearance)]
+        public bool Ripple { get; set; } = true;
+
+        /// <summary>
+        /// The URL to navigate to when this item is clicked.
+        /// </summary>
+        [Parameter]
+        [Category(CategoryTypes.General.ClickAction)]
+        public string? Href { get; set; }
+
+        /// <summary>
+        /// Performs a full page load during navigation.
+        /// </summary>
+        /// <remarks>
+        /// Defaults to <c>false</c>. When <c>true</c>, client-side routing is bypassed and the browser is forced to load the new page from the server.
+        /// </remarks>
+        [Parameter]
+        [Category(CategoryTypes.General.ClickAction)]
+        public bool ForceLoad { get; set; }
+
+        /// <summary>
+        /// The content within this item.
+        /// </summary>
+        [Parameter]
+        [Category(CategoryTypes.General.Behavior)]
+        public RenderFragment? ChildContent { get; set; }
+
+        /// <summary>
+        /// Occurs when the item has been clicked.
+        /// </summary>
+        /// <remarks>
+        /// This event only occurs when the <see cref="Href"/> property is not set.
+        /// </remarks>
+        [Parameter]
+        public EventCallback<MouseEventArgs> OnClick { get; set; }
+
         protected Task HandleNavigation()
         {
             if (!Disabled && NavigationEventReceiver != null)
@@ -79,6 +133,23 @@ namespace MudBlazor
             }
 
             return Task.CompletedTask;
+        }
+
+        protected async Task OnClickHandler(MouseEventArgs ev)
+        {
+            if (Disabled)
+            {
+                return;
+            }
+
+            if (Href is not null)
+            {
+                UriHelper.NavigateTo(Href, ForceLoad);
+            }
+            else
+            {
+                await OnClick.InvokeAsync(ev);
+            }
         }
     }
 }
