+++ src/MudBlazor/Components/Select/MudSelectItem.razor
@@ -1,6 +1,6 @@
 ﻿@namespace MudBlazor
 @typeparam T
-@inherits MudBaseSelectItem
+@inherits MudComponentBase
 
 @if (!HideContent)
 {
